pyproject.toml
[project]
name = "flora-asset-pipeline"
version = "0.1.0"
description = "Automated botanical data and image pipeline for the Flora iOS app"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.30.0",
    "pgvector>=0.3.6",
    "alembic>=1.14.0",
    "httpx>=0.28.0",
    "beautifulsoup4>=4.12.0",
    "wikipedia-api>=0.7.1",
    "rembg>=2.0.57",
    "pillow>=11.0.0",
    "numpy>=1.26.0",
    "structlog>=24.4.0",
    "opentelemetry-api>=1.29.0",
    "opentelemetry-sdk>=1.29.0",
    "opentelemetry-instrumentation-fastapi>=0.50b0",
    "opentelemetry-exporter-prometheus>=0.50b0",
    "prometheus-client>=0.21.0",
    "boto3>=1.35.0",
    "mlflow>=2.19.0",
    "transformers>=4.47.0",
    "sentencepiece>=0.2.0",
    "sacremoses>=0.1.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-httpx>=0.35.0",
    "respx>=0.21.0",
    "ruff>=0.9.0",
    "mypy>=1.13.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["backend", "cli"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
